Move control files
