+++ src/components/controls/Pan.vue
