+++ src/components/controls/Zoom.vue
